Add a project customization workflow
I added a workflow that runs when a new project is created from the
template repository. The workflow runs steps to customize the new
repository, such as changing references to the template repo.

I also renamed the template quality workflow to static analysis. It
makes the customization easier, and is still applicable to the
template repo.

closes #3